requirements.txt
# Twitter SaaS Idea Finder - Dependencies

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Scraping - Primary (twscrape)
twscrape>=0.12.0

# Scraping - Backup (Apify)
apify-client>=1.6.0

# Scraping - Fallback (Selenium/Scweet)
scweet>=2.2
selenium>=4.16.0
webdriver-manager>=4.0.1

# Legacy Twitter API (fallback)
tweepy>=4.14.0

# CLI and Output
rich>=13.7.0
click>=8.1.7

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Background Jobs
celery>=5.3.4
flower>=2.0.1
redis>=5.0.1
apscheduler>=3.10.4

# HTTP Requests
httpx>=0.26.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
lxml>=5.1.0

# Utils
python-dateutil>=2.8.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Optional: OCR (uncomment if needed)
# pytesseract>=0.3.10
# pillow>=10.0.0

# Optional: Traffic Analysis
# similarweb>=1.0.0
